Sched: cooperative sleep/wake with stack canaries (rung 2)
Rung-2 scheduler: threads can sleep on a deadline and the scheduler
runs other threads (or the CPU idles via wfi) until they're due. The
prior busy-wait sleep (kernel::timer::sleep_ms) becomes
sched::sleep — backed by a Sleeping(u64) state variant in the TCB.

State machine: Avail / Ready / Running / Sleeping(deadline). The
data-carrying enum keeps the wakeup deadline coupled to the variant,
so invalid combinations (e.g., Ready + Some(deadline)) are
unrepresentable.

Reschedule unification: yield_now and sleep_until both call a single
private reschedule(new_state) that:
- sets the current thread's state to new_state up front,
- runs a wake-check (Sleeping → Ready when deadline <= now),
- searches 1..=THREADS_MAX from curr+1 (cyclic, includes curr),
- if found a Ready thread (curr or another), switches or self-resumes,
- if none found, drops the lock, wfi, reacquires, retries.

Same loop handles both the rung-1 yield case (curr stays Ready, found
on first iteration) and the new sleep-until-deadline case (curr
re-promotes itself when the wake-check fires after enough wfi
iterations).

Stack canaries: each per-thread stack gets 0xDEADBEEF written at
offset 0..4 in spawn (and bootstrap, so the boot thread's slot
satisfies the assertion vacuously). reschedule asserts the canary
on entry; overflow is caught at switch time with a clear message
rather than manifesting as illegal-instruction at a random address.
THREAD_STACK_SIZE bumped 1 KiB → 4 KiB; print! formatting plus
scheduler frames was tight at 1 KiB and reliably crashed with three
threads.

Timer: TICKS upgraded from AtomicUsize (32-bit, 49-day wrap) to a
two-AtomicU32 pair (TICKS_L + TICKS_H) with Lamport-style read
protocol. ticks_ms() now returns u64. RV32IMAC has no 64-bit atomic
ops, so the simple paired counter with retry-if-high-changed is the
right tradeoff. Callers in virtio and shell/keyboard updated to u64.

arch::wait_for_interrupt: a single wfi instruction returning
normally. Returns () not !; the loop belongs in the scheduler, not
the arch primitive.

Tests + benchmark in coop_rr.rs:
- yield_makes_progress: spawn a partner that bumps a counter, do 10
  yields, assert it increased.
- sleep_blocks_for_duration: measure sched::sleep(100) against
  ticks_ms; tolerance 95..=200 ms.
- sleep_until_past_returns_quickly: deadline 10 ms in the past
  returns within 5 ms (wake-check immediate-fire path).
- bench_yield_cycles: 1000 yield_now calls timed via Clint::mtime,
  prints cycles/call. No assertion.

Tested with three threads (B/A/C at 500/700/300 ms periods)
producing the expected irregular interleaving that proves
independent scheduling.

+++ src/drivers/virtio/mod.rs
@@ -253,7 +253,7 @@ impl VirtioBlkDev {
     }
 }
 
-const IO_TIMEOUT_MS: usize = 1_000;
+const IO_TIMEOUT_MS: u64 = 1_000;
 
 // Lock held when IO is in progress (interrupts enabled)
 static IO_IN_PROGRESS: SpinLock<()> = SpinLock::new(());
